feat add game
game meg nincs keszen,
dashboardrol atdobjon a mapra,
mapon csak a publikus kalandokat jelenitse meg.
kaland profil es inditas

+++ geoszabaduloszoba-backend/src/main/java/com/cityscape/geoszabaduloszobabackend/model/dto/AdventureProfileDTO.java
@@ -10,7 +10,7 @@ public class AdventureProfileDTO {
     private String description;
 
     private String averageTime;
-    private Integer distanceInMeters;
+    private Double distanceInMeters;
     private String difficulty;
     private String creatorName;
